PixelPaint: Move files about layers into directory 'Layer'
