+++ Userland/Applications/PixelPaint/CMakeLists.txt
@@ -12,7 +12,6 @@ compile_gml(ResizeImageDialog.gml ResizeImageDialogGML.h resize_image_dialog_gml
 compile_gml(LevelsDialog.gml LevelsDialogGML.h levels_dialog_gml)
 
 set(SOURCES
-    BitmapLayer.cpp
     CreateNewImageDialog.cpp
     CreateNewLayerDialog.cpp
     EditGuideDialog.cpp
@@ -38,7 +37,8 @@ set(SOURCES
     IconBag.cpp
     Image.cpp
     ImageEditor.cpp
-    Layer.cpp
+    Layers/BitmapLayer.cpp
+    Layers/Layer.cpp
     LayerListWidget.cpp
     LayerPropertiesWidget.cpp
     LevelsDialogGML.h
