+++ Userland/Applications/PixelPaint/LayerListWidget.cpp
@@ -9,7 +9,7 @@
 #include "LayerListWidget.h"
 #include "Image.h"
 #include "ImageEditor.h"
-#include "Layer.h"
+#include "Layers/Layer.h"
 #include <LibGUI/Painter.h>
 #include <LibGfx/Palette.h>
