+++ Userland/Applications/PixelPaint/Layers/BitmapLayer.h
