+++ Userland/Applications/PixelPaint/LevelsDialog.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #include "ImageEditor.h"
-#include "Layer.h"
+#include "Layers/Layer.h"
 #include <LibGUI/Dialog.h>
 
 namespace PixelPaint {
