+++ Userland/Applications/PixelPaint/MainWidget.h
@@ -11,9 +11,9 @@
 #include "IconBag.h"
 #include "Image.h"
 #include "ImageEditor.h"
-#include "Layer.h"
 #include "LayerListWidget.h"
 #include "LayerPropertiesWidget.h"
+#include "Layers/Layer.h"
 #include "PaletteWidget.h"
 #include "ProjectLoader.h"
 #include "ToolPropertiesWidget.h"
