+++ Userland/Applications/PixelPaint/Tools/BucketTool.cpp
@@ -8,7 +8,7 @@
 
 #include "BucketTool.h"
 #include "../ImageEditor.h"
-#include "../Layer.h"
+#include "../Layers/Layer.h"
 #include <AK/HashTable.h>
 #include <AK/Queue.h>
 #include <LibGUI/BoxLayout.h>
