+++ Userland/Applications/PixelPaint/Tools/CloneTool.cpp
@@ -6,7 +6,7 @@
 
 #include "CloneTool.h"
 #include "../ImageEditor.h"
-#include "../Layer.h"
+#include "../Layers/Layer.h"
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/Label.h>
 #include <LibGUI/Menu.h>
