+++ Userland/Applications/PixelPaint/Tools/EllipseTool.cpp
@@ -8,7 +8,7 @@
 
 #include "EllipseTool.h"
 #include "../ImageEditor.h"
-#include "../Layer.h"
+#include "../Layers/Layer.h"
 #include <LibGUI/Action.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/CheckBox.h>
