+++ Userland/Applications/PixelPaint/Tools/EraseTool.cpp
@@ -8,7 +8,7 @@
 
 #include "EraseTool.h"
 #include "../ImageEditor.h"
-#include "../Layer.h"
+#include "../Layers/Layer.h"
 #include <LibGUI/Action.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/CheckBox.h>
