+++ Userland/Applications/PixelPaint/Tools/LineTool.cpp
@@ -9,7 +9,7 @@
 
 #include "LineTool.h"
 #include "../ImageEditor.h"
-#include "../Layer.h"
+#include "../Layers/Layer.h"
 #include <AK/Math.h>
 #include <LibGUI/Action.h>
 #include <LibGUI/BoxLayout.h>
