+++ Userland/Applications/PixelPaint/Tools/PenTool.cpp
@@ -8,7 +8,7 @@
 
 #include "PenTool.h"
 #include "../ImageEditor.h"
-#include "../Layer.h"
+#include "../Layers/Layer.h"
 #include <LibGUI/Action.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/Label.h>
