+++ Userland/Applications/PixelPaint/Tools/RectangleSelectTool.cpp
@@ -7,7 +7,7 @@
 
 #include "RectangleSelectTool.h"
 #include "../ImageEditor.h"
-#include "../Layer.h"
+#include "../Layers/Layer.h"
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/Button.h>
 #include <LibGUI/ComboBox.h>
