+++ Userland/Applications/PixelPaint/Tools/RectangleTool.cpp
@@ -9,7 +9,7 @@
 
 #include "RectangleTool.h"
 #include "../ImageEditor.h"
-#include "../Layer.h"
+#include "../Layers/Layer.h"
 #include <LibGUI/Action.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/CheckBox.h>
